backend/requirements.txt
# Core
fastapi==0.115.6
uvicorn[standard]==0.34.0
pydantic==2.10.4
pydantic-settings==2.7.1

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.18

# HTTP & CORS
httpx==0.28.1
python-dotenv==1.0.1

# Rate Limiting (in-memory fallback when no Redis)
slowapi==0.1.9

# Dev
pytest==8.3.4
pytest-asyncio==0.25.0
